pless: Rename emaildelivery/service to emaildelivery/services
