+++ supertokens_python/recipe/passwordless/utils.py
@@ -32,7 +32,7 @@ if TYPE_CHECKING:
 from re import fullmatch
 
 from phonenumbers import is_valid_number, parse  # type: ignore
-from supertokens_python.recipe.passwordless.emaildelivery.service.backward_compatibility import \
+from supertokens_python.recipe.passwordless.emaildelivery.services.backward_compatibility import \
     BackwardCompatibilityService
